mod util.h

+++ Lecture06/ex07/util.h
@@ -3,7 +3,8 @@
 typedef unsigned char byte;
 typedef unsigned short u16bit;
 typedef unsigned int u32bit;
-typedef uint64_t u64bit;
+typedef unsigned long long u64bit;
+// typedef uint64_t u64bit;
 
 typedef signed int s32bit;
